requirements.txt
# FunctionGemma Fine-tuning for Home Assistant
# GPU Training Dependencies

# Core ML
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.25.0
datasets>=2.16.0

# Efficient fine-tuning
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0

# Home Assistant API
aiohttp>=3.9.0
websockets>=12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0

# Evaluation
evaluate>=0.4.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
